cincinnati: evaluate dead-end status for booted OS
This adds logic to evaluate and record whether the booted deployment
is a dead-end release from the point of view of auto-updates graph
(i.e. it cannot further auto-update).
As a dead-end release requires manual intervention to get out of its
stuck state, this piece of information has to be bubbled-up to
monitoring systems (e.g. via metrics).