rpm-ostree/status: detect and handle staged deployments
This augments rpm-ostree status querying logic in order to detected
staged-but-not-finalized deployments, i.e. those that haven't really
been persisted locally in the past.
Zincati uses this new logic in order to distinguish between previous
deployments (e.g. in case of rollbacks) and temporarily staged ones
(e.g. in case of manual interventions or half-applied updates). The
latter ones can be considered as future update targets, while the
former ones should not.