cincinnati: Add metric for number of ignored update targets
This differs from the total number of deployments that are found
at startup in that ignored update targets are targets that are
actually found to be available releases during a check for udpates.

This metric, in most cases, should indicate whether a node has
performed a rollback after an update.