update_agent: delay reboot if users logged in
Allow update finalizations (reboots) to be postponed a number of times
for an amount of time if it is detected that there are users (with a tty)
currently logged in on the system.
Add a counter for number of postponements remaining to the UpdateStaged
state of the update agent; if there are no allowed postponements remaining,
proceed with the finalization, disregarding active users.