cli: Check if run as `zincati` user
Error out and log warning message if not run by the `zincati` user
for the `agent` and `deadend-motd` subcommands, since the update
agent should only be running as the `zincati` user and `deadend-motd`
should not be used externally.
The binary is currently already placed in `/usr/libexec` in Fedora,
but doing this for extra safety and clarity.

Closes: https://github.com/coreos/zincati/issues/95