cli: make check for user more flexible
Add a `user` argument to `ensure_user()` so we can ensure different
users flexibly. Specifically, for the `agent` subcommand, we want
to ensure that only the `zincati` user can run it. But for the
`deadend-motd` subcommand, we want to make sure that it is run by
`root`, otherwise, the creation of the MOTD will fail anyway.

Note that `deadend-motd` is still only meant to be called internally by
the `agent` process started by `zincati.service` as the `zincati`
user; but in order to write the MOTD, the `agent` process `pkexec`s
as the `root` user before calling the `deadend-motd` subcommand. This
is the reason that we do not check that user is `zincati`, but `root`,
instead.

Addresses: https://github.com/coreos/zincati/pull/495#discussion_r593107325