Support time zones for periodic strategy
Include a single `time_zone` field (that defaults to UTC if not
specified) in the `periodic` strategy's configuration. Whenever we
need to check whether we're currently in an allowed reboot window,
we convert the current naive time to the naive time under the
specified time zone, and check whether we are currently in a reboot
window.

Under this implementation, we keep invariant clock time for reboots,
but do NOT keep invariant reboot window length; thus, in some cases,
reboot windows can be lengthened, shortened, or skipped entirely.