update_agent: promote some log messages to INFO level
It should be safe to promote to INFO level the message that Zincati
has successfully staged a deployment and the message that Zincati has
reached its "steady" state, without the risk of being overly verbose.
This will help admins get a better idea of which
stage in the state machine Zincati is in.