update_agent: put agent state behind a RwLock
We would like to put most of the "writable" parts of the update
agent behind a RwLock in order to make sure that, when the
update_agent actor receives multiple messages, it sequentially
processes the messages that would like to write/update the agent
state to avoid races.

In order to do this, we explicitly separate out a `state` field
and an `info` field in the `UpdateAgent` struct, where the `state`
field holds the `UpdateAgentState` that is behind a RwLock.
Though nothing enforces this, all the fields (except for
`state_changed`) in `UpdateAgentInfo` should be read-only, and
needs not be behind any lock.

During each tick, because we are passing the asynchronously acquired
lock guard of the agent state across futures, we can no longer make
use of Actix's `ResponseActFuture` (which contains the actor state and
context). We now use the newer `async/await` syntax where we can to
make it easier to move data across futures and generally make the code
easier to read. For context, the reason we previously stuck to the less
ergonomic future combinator (`.then()`) syntax was because we wanted to
make use of `ResponseActFuture`.