update_agent: remove unused `Result`s
Previously, the `tick_` functions returned a `Result`, but only two
of these functions would ever return the `Err` variant. Even in the
case where the returned `Result` was of the `Err` variant, the caller
does not really do anything with it.

Since we already do a good amount of error logging in the inner
functions, refactor the `tick_` functions to return `()` instead.