update_agent: split state_changed out of UpdateAgentInfo
We would like for everything under `UpdateAgentInfo` to be read-only.
Conceptually, it would make sense to put the `state_changed` field
behind the same lock as `UpdateAgentState`, but `UpdateAgentState` is
behind a `RwLock`, which doesn't allow for reads when it is acquired for
writing.
Get around this by manually ensuring that we only write to
`state_changed` when the `RwLock` that `UpdateAgentState` is behind is
acquired for writing. This allows us to ensure that the `RefCell` that
`state_changed` is behind will not panic, and also allows for reads even
when the `UpdateAgentState`'s `RwLock` is acquired for writing.