update_agent: use `Cell` for `state_changed` field
The previous incorrect usage of `RefCell` allowed for panics at runtime
if there were consumers.
`Cell` should be sufficient here, since the data that we are storing
is a simple timestamp. We do not need to use an atomic type because the
update_agent actor runs on a single thread on the SystemArbiter.