update_agent: record a denylist in the state of update agent
Previously, we had a read-only `UpdateAgentInfo` and a read-write
`UpdateAgentState` that was an enum representing the state of the
update agent in the state machine.
Rename this enum `UpdateAgentMachineState` and create a new struct
`UpdateAgentState` which encapsulates all the non-atomic read-write
data associated with the update agent. All this non-atomic read-write
data will be behind the RwLock.
Add a `denylist` field to this new read-write struct. During
initialization, save local finalized deployments to this denylist.
When fetching for new updates from Cincinnati, pass this denylist to
Cincinnati.