update_agent: log progress while waiting for the end of user sessions
This adds some logging to each relevant step in the logic which waits
for the end of all interactive sessions.
Zincati currently emits some details in real time to the consoles
and updates the service status, but these additional log messages
make it easier to get an historical view of the events (from the
journal).