tests/kola: re-arrange shared test helpers
This reworks the symlinked helpers in order to fix CI environment.
There was a change in some external components (which I couldn't
exactly pinpoint) and in resulted in an empty /var/opt under our
existing setup.
Moving the symlink from the whole 'data' directory to the
individual file seems to fix that. Not great, but it unstucks CI.