Remove requirement for `coreos-assembler.basearch`
This is aiming to increase compatibility of zincati
with systems that aren't derived from Fedora CoreOS (e.g. IoT,
desktop systems, etc.)

Additionally, even in the FCOS case, when one derives a container
image from FCOS, the final build isn't generated by coreos-assembler
and hence the ostree commit won't have this metadata.

In the end - I can't think of a use case for even trying
to have zincati managing an image for a different architecture
than the one it's compiled for.

So, copy the code I wrote for oci-spec-rs which maps
from the Rust architecture to the GNU one.  xref
https://github.com/containers/oci-spec-rs/commit/52e450e631a212769df4dcd5b7904d942d246d24