Simplify read paths
Use the `std::fs` functions to read to a string or `Vec<u8>` directly;
it's more efficient and less code.

Also in one place use `.map(BufReader::new)` which I find more...
readable.

Not prep for anything, just a drive-by change since I was
reading the code for other reasons.