update_agent: handle new output format for loginctl
In systemd 254 the `loginctl list-sessions --json` output was modified slightly:

Before
```
[{"session":"4","uid":1000,"user":"core","seat":"","tty":""}]
```

After
```
[{"session":"17","uid":1000,"user":"core","seat":"","tty":"n/a","state":"active","idle":false,"since":null}]
```

Notice the change from `""` to `"n/a"` for the `tty` field.

This change probably happened in one of:

- https://github.com/systemd/systemd/pull/27769
- https://github.com/systemd/systemd/pull/27740
- https://github.com/systemd/systemd/pull/27606

Let's update the code to handle the new case and rename the function to
be a little more appropriate since the use case for it is narrow.

Fixes https://github.com/coreos/fedora-coreos-tracker/issues/1547