update_agent: add hidden knob to override interactive session check
An issue as old as ~time~Fedora CoreOS is that when you're testing
something related to updates, you want Zincati to reboot but because
`cosa run` gives you an interactive session and you can't *not* have a
session without losing the VM, you're stuck waiting for the 10 minutes
grace period to expire.

Just add a hidden knob for this to make that scenario less painful.

The API is pretty much: when you're ready to have Zincati reboot, just
`touch /run/zincati/override-interactive-check`.