update_agent: support updating with OCI graph
Local deployements are now unserialized with a enum being either
`Checksum` or `Pullspec`.
This will cue the cincinnati client to request the OCI graph in the
pullspec case. [1]

The custom-origin-description, if set, is passed along the rebase
call, and the custom-url is updated to match the new pullspec. [2]

Finally, update the polkit policy to allow Zincati
to do a rebase operation.

[1] Requires https://github.com/coreos/fedora-coreos-cincinnati/pull/99
and https://github.com/coreos/rpm-ostree/pull/5120