dist/polkit-1/rules.d: fix boolean logic
Commit 28a43aa2 introduces a regression in the polkit rules by
misplacing a parenthesis. The && operator has higher precedence than
||, so resolves first. The end result is that while the `cleanup` action
was properly gated to the `zincati` user, the `deploy`, `rebase`, and
`finalize-deployment` were not. And therefore, any unprivileged user
with access to the host D-Bus was capable of executing those actions.

Instead of fixing the misplaced parenthesis, just reflow this so that we
make the check for the `zincati` user entirely separate.

Fixes: GHSA-w6fv-6gcc-x825