Add new `zincati-update-now` command
This addresses the gap identified in
https://github.com/coreos/zincati/issues/498 in a blunt way by setting
some runtime config knobs so Zincati in effect immediately downloads,
deploys, and reboots the system.

The motivation to at least have this for now is
https://github.com/coreos/zincati/pull/1241, which will purposely break
`rpm-ostree upgrade --bypass-driver`. But clearly the approach here is
applicable regardless of whether the node has switched to OCI.

In the future, this may either get replaced by a proper `zincatictl`
entrypoint, or by better integration into the system's update tooling
(e.g. bootc), so that `bootc upgrade` does the right thing.