Fixed minute format in periodic update log
The log message reporting the time of the next update window for the `periodic` strategy does not format the minute nicely. For instance, an update scheduled for `2:00` appears instead as `2:0` in the log, which may make new users think something is configured wrong. This commit simply zero-pads the minute to ensure two digits are always displayed.