.cci.Jenkinsfile: Separate stages for Build/Test/Install
`make` will run the unit tests too so we have been running the
unit tests for some time and just needed to drop the commented
out lines with the `// XXX`.

Let's make it more clear that we are running the unit tests by
separating out the Build/Test/Install into separate stages.